Moved RabbitMqOptions to Contracts project
